Nuke frontend_input stuff and merge it with event loop
It was never really necessary, and every backend just uses a generic
"get all events" thing, not "get all inputs and only inputs" thing.

This was only a good idea if you could have separate event and input
loops, but this was never the case, and is unlikely to be the case for
any other event loop implementation. Even if it is, it can be handled
appropriately in the given event loop.