Bounds-check vram_{read|write}.  Let me explain.
I don't know exactly WHAT triggers this bug, but Donkey Kong was able to
once under some random condition, and Coverity is also identifying this
as an error with certain values of reg.  I don't know where exactly this
is called with values that are not in that boundary, but this is a .003x
speed penalty after 180 seconds on my 2.4GHz i5 so I don't see it being
a horrible degradation of performance to avoid a segmentation fault.  If
anyone disagrees, feel free to revert this change.