Oh this is a FUN one!  Fix SDL2 build on Mac OS X.
SDL2, on Mac OS X only, does some really bizarre fucko shit with bools.
Like, there's no other way to describe it.

This fixes it with a stupid yet hilarious hack involving checking if we
have mach/clock.h in a module we don't use it in just to check if OS X
is present, since we don't currently HAVE a platform definition for OS
X, it's "just POSIX".

It Just Werks®, right?